+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/GoogleMapOptionsSink.java
@@ -64,4 +64,6 @@ interface GoogleMapOptionsSink {
   void setInitialGroundOverlays(@NonNull List<PlatformGroundOverlay> initialGroundOverlays);
 
   void setMapStyle(@Nullable String style);
+
+  void setMapColorScheme(@Nullable Integer colorScheme);
 }