+++ packages/google_maps_flutter/google_maps_flutter_android/pigeons/messages.dart
@@ -483,6 +483,9 @@ class PlatformMapViewCreationParams {
 
 enum PlatformMarkerType { marker, advancedMarker }
 
+/// Pigeon equivalent of MapColorScheme.
+enum PlatformMapColorScheme { light, dark, followSystem }
+
 /// Pigeon equivalent of MapConfiguration.
 class PlatformMapConfiguration {
   PlatformMapConfiguration({
@@ -507,6 +510,7 @@ class PlatformMapConfiguration {
     required this.markerType,
     required this.mapId,
     required this.style,
+    required this.colorScheme,
   });
 
   final bool? compassEnabled;
@@ -530,6 +534,7 @@ class PlatformMapConfiguration {
   final PlatformMarkerType markerType;
   final String? mapId;
   final String? style;
+  final PlatformMapColorScheme? colorScheme;
 }
 
 /// Pigeon representation of an x,y coordinate.